fix parsing generating huffman tables

+++ src/jpg/dec/scan.c
@@ -20,16 +20,16 @@
 #include <stdio.h>
 #include <math.h>
 
-  uint32_t unzig[64] = {
-    0,  1,  8,  16, 9,  2,  3,  10,
-    17, 24, 32, 25, 18, 11, 4,  5,
-    12, 19, 26, 33, 40, 48, 41, 34,
-    27, 20, 13, 6,  7,  14, 21, 28,
-    35, 42, 49, 56, 57, 50, 43, 36,
-    29, 22, 15, 23, 30, 37, 44, 51,
-    58, 59, 52, 45, 38, 31, 39, 46,
-    53, 60, 61, 54, 47, 55, 62, 63
-  };
+uint32_t unzig[64] = {
+  0,  1,  8,  16, 9,  2,  3,  10,
+  17, 24, 32, 25, 18, 11, 4,  5,
+  12, 19, 26, 33, 40, 48, 41, 34,
+  27, 20, 13, 6,  7,  14, 21, 28,
+  35, 42, 49, 56, 57, 50, 43, 36,
+  29, 22, 15, 23, 30, 37, 44, 51,
+  58, 59, 52, 45, 38, 31, 39, 46,
+  53, 60, 61, 54, 47, 55, 62, 63
+};
 
 IM_HIDE
 ImByte*
@@ -45,7 +45,7 @@ jpg_decode_dc(ImJpeg    * __restrict jpg,
               ImScan    * __restrict scan,
               ImHuffTbl * __restrict huff,
               int16_t   * __restrict zz) {
-  int32_t t, diff;
+  int16_t t, diff;
 
   t     = jpg_decode(scan, huff);
   diff  = jpg_receive(scan, huff, t);
@@ -60,7 +60,7 @@ jpg_decode_ac(ImJpeg    * __restrict jpg,
               ImScan    * __restrict scan,
               ImHuffTbl * __restrict huff,
               int16_t   * __restrict zz) {
-  int32_t k, rs, ssss, r;
+  int16_t k, rs, ssss, r;
 
   k = 1;
